feat: store array length as `u32`

+++ compiler/noirc_evaluator/src/acir/acir_variable.rs
@@ -92,7 +92,7 @@ impl<'a> From<&'a SsaType> for AcirType {
             SsaType::Numeric(numeric_type) => AcirType::NumericType(*numeric_type),
             SsaType::Array(elements, size) => {
                 let elements = elements.iter().map(|e| e.into()).collect();
-                AcirType::Array(elements, *size)
+                AcirType::Array(elements, *size as usize)
             }
             _ => unreachable!("The type {value} cannot be represented in ACIR"),
         }
